most of things are done, need to solve error reading file and casting

+++ ex1/common.h
@@ -3,6 +3,8 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <stdlib.h>
+#include <string.h>
+#include <assert.h>
 
 #define BUFF_SIZE 1024
 
@@ -15,7 +17,7 @@ struct Records
 };
 
 void create_array(const char* file_path,struct Records *array,int size);
-int n_row(const char *file_path);
+size_t n_row(const char *file_path);
 void merge_binary_insertion_sort(void *base,size_t nitems,size_t size,size_t k,int (*compare)(const void*,const void *));
 void binary_insertion_sort(void *array,int left,int right,int (*compare)(const void*,const void *));
